Fix crash when pressing delete during recording
Crash would happen if del or X was pressed with the mouse cursor
in the sequencer area, because the temp visual feedback color strip
gets deleted while this add-on tries to update its size.

It's unclear what is the goal of pressing delete.
Maybe to stop recording, or maybe just an accident?
My solution is for the visual feedback strip not to be selected,
so delete won't affect it. Users can use the sequencer as normal,
while recording including to delete strips.

If the feedback strip does get intentionally delete, the recording
is canceled, but there is still a chance it may crash. (race?)