Specify file buffer size and flush data to disk as soon as possible
Without specifying the buffer size, ffmpeg will try to determine the
best buffersize and will usually pick a big value to minimize how often
it writes to disk.

By specifying a small value and forcing on "flush_packets", ffmpeg will
write the recorded data back to disk ASAP.

This makes it so there is very little data loss when stopping the
recording.

Closes #4