Add atunc for sound recording on macOS
Atunc is a new utility to record sound on macOS, instead of ffmpeg.
This fixes the glitchy sound and comes bundled in the add-on so no
external dependencies are required.

In the atunc directory we ship the code-signed archive that is
extracted by the add-on on activation, along with the source code of
the app.