Snap end of recorded strip to playhead
Due to recording latency, audio strips might end up being shorter than
the distance travelled by the playhead from the moment we hit record.

This sometimes happens because we currently lose several frames worth
of audio when starting a recording. This is due to latency of ffmpeg
or atunc. We accept that.

In order to make the clip sync-up with the timeline, we need to align
its end to the frame where we stopped the recording.